--story=1002524 --user=nmsl 作为买家 我想查看我的发起的所有售后服务 https://www.tapd.cn/54028298/s/1000816

+++ other-service/src/main/java/cn/edu/xmu/oomall/other/model/vo/Aftersale/AftersaleRetVo.java
@@ -14,7 +14,7 @@ public class AftersaleRetVo {
     private Long customerId;
     private Long shopId;
     private String serviceSn;
-    private Integer type;
+    private Byte type;
     private String reason;
     private String conclusion;
     private Long refund;
